requirements/dev.txt
#
# This file is autogenerated by pyproject2conda
# with the following command:
#
#     $ pyproject2conda project --verbose
#
# You should not manually edit this file.
# Instead edit the corresponding pyproject.toml file.
#
bottleneck
importlib_resources
ipykernel
ipython
ipywidgets
joblib
lazy_loader
matplotlib
module-utilities>=0.9.0
mypy>=1.6.1
nbval
nox
numpy
pandas-stubs
pytest
pytest-accept
pytest-cov
pytest-sugar
pytest-xdist
pytype;python_version<"3.11"
scikit-image>=0.21
scipy
tqdm
types-tqdm
xarray
